Externalized FieldArray component
